Adventure UI updates (#41)
* Portfolio selection in Adventure Creation/Update page now shows a message informing the user if they have no portfolios.

* Began UI overhaul for Adventure screens.

* Delete button confirmation has been changed to a popup.

* Updated UI of DropDownTab and Single Select list

* Updated UI of Multi Selection list. Completed update of Create/Update adventure screen.

* Updated UI of Section Creation and Update Screens. Updated Delete and Save Buttons in Adventure Creation and Update to be icons.

* Updated UI of Adventure View.

+++ app/src/main/java/com/example/active_portfolio_mobile/composables/adventure/DeleteButtonWithConfirm.kt
@@ -1,16 +1,22 @@
 package com.example.active_portfolio_mobile.composables.adventure
 
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.material3.Button
-import androidx.compose.material3.Card
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 
 /**
  * A button for performing a deletion operation, which shows a confirmation pop-up when pressed.
@@ -22,27 +28,37 @@ fun DeleteButtonWithConfirm(deleteFunc: () -> Unit) {
     var confirmDelete by remember { mutableStateOf(false) }
 
     Column(horizontalAlignment = Alignment.CenterHorizontally) {
-        Button(onClick = {
+        IconButton(onClick = {
             confirmDelete = true
         }) {
-            Text("Delete")
+            Icon(imageVector = Icons.Filled.Delete,
+                contentDescription = "Delete Button",
+                modifier = Modifier.size(30.dp)
+            )
         }
-        if (confirmDelete) {
-            Card {
-                Text("Are you sure? Deletion is permanent")
-            }
-            Row {
-                Button(onClick = {
-                    deleteFunc()
-                }) {
-                    Text("Yes, Delete")
+        if(confirmDelete) {
+            AlertDialog(
+                onDismissRequest = { confirmDelete = false },
+                title = {
+                    Text(text = "Confirm Deletion")
+                },
+                text = {
+                    Text("Are you sure? Deletion is permanent")
+                },
+                confirmButton = {
+                    TextButton(onClick = {
+                        deleteFunc()
+                        confirmDelete = false
+                    }) {
+                        Text("Yes, Delete")
+                    }
+                },
+                dismissButton = {
+                    TextButton(onClick = { confirmDelete = false }) {
+                        Text("Cancel")
+                    }
                 }
-                Button(onClick = {
-                    confirmDelete = false
-                }) {
-                    Text("Cancel")
-                }
-            }
+            )
         }
     }
 }
